# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Codalla AI Coding Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Success Criteria ────────────────────────────────────────────────────────
...
  /projects/{projectId}/memory:
    get:
      operationId: listMemoryNotes
      tags: [memory]
      summary: List memory notes for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of memory notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryNote"
        "404":
          description: Project not found
    post:
      operationId: createMemoryNote
      tags: [memory]
      summary: Create a memory note
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryNoteInput"
      responses:
        "201":
          description: Created memory note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryNote"
        "404":
          description: Project not found
  /projects/{projectId}/memory/{noteId}:
    get:
      operationId: getMemoryNote
      tags: [memory]
      summary: Get a memory note by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: noteId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Memory note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryNote"
        "404":
          description: Not found
    patch:
      operationId: updateMemoryNote
      tags: [memory]
      summary: Update a memory note
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: noteId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryNoteUpdate"
      responses:
        "200":
          description: Updated memory note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryNote"
        "404":
          description: Not found
    delete:
      operationId: deleteMemoryNote
      tags: [memory]
      summary: Delete a memory note
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: noteId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── Filesystem ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, localPath, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        isShared:
          type: boolean
        isOwner:
          type: boolean
        localPath:
          type: string
        gitRemoteUrl:
          type: ["string", "null"]
        currentBranch:
          type: string
        lastSynced:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        story:
          type: ["string", "null"]
        target:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        gitRemoteUrl:
          type: string
        localPath:
          type: string
          description: Attach an existing folder on the server (must be under a CODALLA_DATA_ROOTS entry) instead of creating a new project directory
        story:
          type: string
        target:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isShared:
          type: boolean
        gitRemoteUrl:
          type: string
        currentBranch:
          type: string
        story:
          type: string
        target:
          type: string

    # ─── Success Criteria ──────────────────────────────────────────────────────
...
    MemoryNote:
      type: object
      required: [id, projectId, title, content, createdAt, updatedAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    MemoryNoteInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        content:
          type: string
    MemoryNoteUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string

    # ─── Apply File ────────────────────────────────────────────────────────────
...